os/board/rtl8730e: Fix Peripheral related Build Warning and SVACE
- Deleted the unused API and sources file that caused Build Warning and SVACE
Build Warning
- Added return for the function does return
- Removed the unused variable
- Performed the correct casting conversion
- Solved the shadowed declaration without use the same variable name for the global and local
- Added the define/undefine condition checking for the variable
SVACE
- DREF_OF_NULL.ASSIGN : Performed NULL checking
- SIMILAR_BRANCHES : Removed the idetical branches and keep one only
- NO_CAST.INTEGER_DIVISION & INTEGER_OVERFLOW : Performed the correct casting conversion
- Added the comment for unable solve UNREACHEABLE_CODE

+++ os/board/rtl8730e/src/component/mbed/targets/hal/rtl8730e/us_ticker.c
@@ -63,7 +63,7 @@ uint32_t us_ticker_read(void)
 	uint64_t us_tick;
 
 	tick_cnt = SYSTIMER_TickGet(); //up counter
-	us_tick = tick_cnt * (1000000 / 32768);
+	us_tick = (uint64_t)tick_cnt * (1000000 / 32768);
 
 	return ((uint32_t)us_tick);
 }
